INT fixed

+++ _printf.c
@@ -22,7 +22,7 @@ int _printf(const char *format, ...)
 			else if (format[i] == '%' && format[i + 1] == 's')
 				l = l + print_str(ptr), i = i + 2;
 			else if (format[i] == '%' && (format[i + 1] == 'd' || format[i + 1] == 'i'))
-				l = l + print_dec(ptr), i = i + 2;
+				l = l + (unsigned int)print_dec(ptr), i = i + 2;
 			else if (format[i] == '%' && format[i + 1] == '%')
 				_putchar('%'), i = i + 2, l++;
 			else
